split off promise-queries file
